pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wastream"
version = "1.0.0"
description = "Stremio addon to convert DDL to streams via debrid services"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn",
    "selectolax",
    "httpx",
    "databases",
    "aiosqlite",
    "asyncpg",
    "pydantic-settings",
    "loguru",
    "pycryptodome",
    "bcrypt",
    "psutil",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["wastream*"]
